Refactoring of functions that calculate a motion profile and removal of FTOA
The motion profile calculation was originally coupled with the calibration value store in NVRAM, but PID calibration/validation requires the ability to specify a different range.  The inner functionality was refactors into a separate function.

The ftoa function was needed when PsocCreator didn't support %f in sprintf.  More recent PsocCreator supports %f so there is no need to ftoa and the extra string storage.