Added timeout to CDC IsReady check to prevent blocking
Original implementation had a while loop waiting for CDC IsReady.  This would work if there USB port was connected to a serial port; otherwise, it seemed to block the main loop (not confirmed).  Added a function taking an unsigned interger count that allows checking to terminate.  If there is no serial port attached to the USB port then we don't care about the serial output.